+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- Add Swift name to `resumeExternalUserAgentFlowWithURL:error:` and update hint. (#966)
 - Replace case range with explicit case labels in OIDTokenUtilities. Addresses issue #947. ([#963](https://github.com/openid/AppAuth-iOS/pull/963))
 
 # 2.1.0